+++ service/src/main/java/com/gmail/vpshulgaa/service/impl/OrderServiceImpl.java
@@ -4,7 +4,7 @@ import com.gmail.vpshulgaa.dao.OrderDao;
 import com.gmail.vpshulgaa.dao.entities.Item;
 import com.gmail.vpshulgaa.dao.entities.Order;
 import com.gmail.vpshulgaa.dao.entities.User;
-import com.gmail.vpshulgaa.dao.enums.Status;
+import com.gmail.vpshulgaa.dao.enums.StatusEnum;
 import com.gmail.vpshulgaa.service.ItemService;
 import com.gmail.vpshulgaa.service.OrderService;
 import com.gmail.vpshulgaa.service.UserService;
@@ -13,7 +13,8 @@ import com.gmail.vpshulgaa.service.converter.DtoConverter;
 import com.gmail.vpshulgaa.service.dto.ItemDto;
 import com.gmail.vpshulgaa.service.dto.OrderDto;
 import com.gmail.vpshulgaa.service.dto.UserProfileDto;
-import com.gmail.vpshulgaa.service.util.ServiceUtils;
+import com.gmail.vpshulgaa.service.exception.EntityNotFoundException;
+import com.gmail.vpshulgaa.service.util.CurrentUserUtils;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -27,8 +28,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class OrderServiceImpl implements OrderService {
-    private static final Logger logger = LogManager.getLogger(OrderServiceImpl.class);
 
+    private static final Logger logger = LogManager.getLogger(OrderServiceImpl.class);
     private final OrderDao orderDao;
     private final Converter<OrderDto, Order> orderConverter;
     private final DtoConverter<OrderDto, Order> orderDtoConverter;
@@ -58,26 +59,12 @@ public class OrderServiceImpl implements OrderService {
     @Override
     @Transactional(readOnly = true)
     public OrderDto findOne(Long id) {
-        OrderDto orderDto = null;
-        try {
-            Order order = orderDao.findOne(id);
-            orderDto = orderDtoConverter.toDto(order);
-        } catch (Exception e) {
-            logger.error("Failed to get order", e);
+        Order order = orderDao.findOne(id);
+        if (order != null) {
+            return orderDtoConverter.toDto(order);
+        } else {
+            throw new EntityNotFoundException(Order.class, id);
         }
-        return orderDto;
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public List<OrderDto> findAll() {
-        List<OrderDto> orders = new ArrayList<>();
-        try {
-            orders = orderDtoConverter.toDtoList(orderDao.findAll());
-        } catch (Exception e) {
-            logger.error("Failed to get orders", e);
-        }
-        return orders;
     }
 
     @Override
@@ -85,12 +72,12 @@ public class OrderServiceImpl implements OrderService {
     public OrderDto create(OrderDto orderDto, Long itemId) {
         try {
             Order order = orderConverter.toEntity(orderDto);
-            Long userId = ServiceUtils.getPrincipal().getId();
+            Long userId = CurrentUserUtils.getPrincipal().getId();
             UserProfileDto userDto = userService.findOne(userId);
             ItemDto itemDto = itemService.findOne(itemId);
             order.setItem(itemConverter.toEntity(itemDto));
             order.setUser(userProfileConverter.toEntity(userDto));
-            order.setStatus(Status.NEW);
+            order.setStatus(StatusEnum.NEW);
             order.setCreated(LocalDateTime.now());
             orderDao.create(order);
             orderDto = orderDtoConverter.toDto(order);
@@ -133,27 +120,11 @@ public class OrderServiceImpl implements OrderService {
     @Override
     @Transactional
     public void deleteById(Long id) {
-        try {
+        if (orderDao.findOne(id) != null) {
             orderDao.deleteById(id);
-        } catch (Exception e) {
-            logger.error("Failed to delete order", e);
+        } else {
+            throw new EntityNotFoundException(Order.class, id);
         }
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public List<OrderDto> findOrdersByUserId(Long userId) {
-        List<OrderDto> ordersDto = new ArrayList<>();
-        List<Order> orders;
-        try {
-            orders = orderDao.findordersByUserId(userId);
-            for (Order order : orders) {
-                ordersDto.add(orderDtoConverter.toDto(order));
-            }
-        } catch (Exception e) {
-            logger.error("Failed to find orders", e);
-        }
-        return ordersDto;
     }
 
     @Override
@@ -173,7 +144,7 @@ public class OrderServiceImpl implements OrderService {
     public Long countOfOrderForUser() {
         Long count = 0L;
         try {
-            Long userId = ServiceUtils.getPrincipal().getId();
+            Long userId = CurrentUserUtils.getPrincipal().getId();
             count = orderDao.countOfOrdersForUser(userId);
         } catch (Exception e) {
             logger.error("Failed to find orders", e);
@@ -203,7 +174,7 @@ public class OrderServiceImpl implements OrderService {
         List<OrderDto> ordersDto = new ArrayList<>();
         List<Order> orders;
         try {
-            Long userId = ServiceUtils.getPrincipal().getId();
+            Long userId = CurrentUserUtils.getPrincipal().getId();
             orders = orderDao.findOrdersByPageForUser(page, maxResults, userId);
             for (Order order : orders) {
                 ordersDto.add(orderDtoConverter.toDto(order));
